sink: split mysql sink into multiple files (#3318)
